projects/01_fyyur/requirements.txt
alembic==1.5.2
Babel==2.9.0
click==7.1.2
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==3.1
infinity==1.5
intervals==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
phonenumbers==8.12.16
psycopg2==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.36.8
validators==0.18.2
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.5
